1. Complete the function that the admin button will only be seen by administrator 2. Complete basic admin UI

+++ app/src/main/java/com/example/connectue/General.java
@@ -2,12 +2,12 @@ package com.example.connectue;
 
 public class General {
 
-//    user role parameters in database
-    public static final int ADMIN = 0;
-    public static final int STUDENT = 1;
-    public static final int GUEST = 2;
+    //    user role parameters in database
+    public static final long ADMIN = 0;
+    public static final long STUDENT = 1;
+    public static final long GUEST = 2;
 
-//    the name of user collection, and the field names within
+    //    the name of user collection, and the field names within
     public static final String USERCOLLECTION = "users";
     public static final String EMAIL = "email";
     public static final String FIRSTNAME = "firstName";
@@ -23,15 +23,15 @@ public class General {
     public static final String USERID = "userId";
     public static final String PHONE = "phone";
 
-    public static boolean isAdmin(int role) {
+    public static boolean isAdmin(long role) {
         return role == ADMIN;
     }
 
-    public static boolean isStudent(int role) {
+    public static boolean isStudent(long role) {
         return role == STUDENT;
     }
 
-    public static boolean isGuest(int role) {
+    public static boolean isGuest(long role) {
         return role != ADMIN && role!= STUDENT;
     }
 }
